Opensea_collection_filter 2.0
The collection filter has been upgraded to accept percentage and decimal values.  As the opensea website included the creator fee in the main page tab. Meaning no extension is needed. However, the value is given in percentages and not decimals. In the fact both data types are being collected in the past and present, it's necessary to work with both data types.