Add files via upload
Inorder to analyze data created by the opensea scraper, a filter is needed. Which filters out faulty data like nan numbers or '---' strings. Moreover, it filters data based on floor price to find out which collections are viable to bid on. Once that's finished, an excel file is made that will be used in the #Opensea Data Graph where the collections are visualized in graphs.