Add files via upload
In order to analyze data created by the opensea scraper, a filter is needed. Which filters out faulty data like nan numbers or '---' strings. Moreover, it filters data based on floor price to find out which collections are viable to bid on. This will be represented in a graph where it shows the potential profit to be made and the ratio between profit and floor price. All collections have royalty profit deduction included. Meaning the graph shows net profit. Based on how often the collection is found in the excel file, it will show in different settings. For example at lane 121, it says between more than 40 and less then 70. Meaning the collections needs to have popped up more than 40 times in the excel file but fewer then 70 times. 

The more often an collection is found, the higher the assumption is of healthy consistent activity. As the data scraper is active for over a month.