Add files via upload
In this file, bidding estimations are made through information from the OpenSea Api. The collection gets loaded in with it's floor price. Which get's put through a filter where a net profit is selected of about 15%. Finally, it gets converted to the 'Bid Links out File'. Where the estimated collection offer get's bid on projects through a Windows Power Automate file.